Fix: Dropdown doesn't work for host with dot
In the problems pan if a host have a dot in his name the dropdown does
not work.
In order to make this work we should not use any dot insight id value
of the pannel-group. I've update the template in order to use
current_problem_id instead of host_name.